+++ JudoPayDotNetTests/Clients/WebPayments/CheckCardTests.cs
@@ -63,7 +63,6 @@ namespace JudoPayDotNetTests.Clients.WebPayments
                     Currency = "GBP",
                     Consumer = new Consumer
                     {
-                        ConsumerToken = "B245SEB",
                         YourConsumerReference = "Consumer1"
                     }
                 }